Handle multiple quoted words in Encodings.unquote_and_convert_to
The function Encodings.unquote_and_convert_to isn't used anywhere in the
current source tree, but it's necessary to correctly handle headers with
non-ASCII, non-UTF-8 character sets under Ruby 1.8.

For example, we can force the unquoting of the 'To' header as UTF-8 using
the following code:

  to_field = mail.header['to']
  Mail::Encodings.unquote_and_convert_to(to_field.value, 'utf-8')

This patch adds support for multiple quoted chunks in the same header.
For example:

  To: =?Shift_JIS?Q?=93=FA=96{=8C=EA=?= <a@example.com>, \
          =?Shift_JIS?Q?=93=FA=96{=8C=EA=?= <b@example.com>

All that was actually needed was a regular expression tweak to prevent
the two quoted chunks from being matched as one giant, malformed chunk.

Note that we still don't make any attempt to correctly deal with headers
that mix different encodings in the same header.  I'm sure these exist
somewhere in the wild, but I haven't seen them yet.

Ideally, all this code should go away and be replaced with a more
robust implementation of field.charset and field.decoded.