Allow bundler to automatically build a gem directly from git
Using bundler (now included with Rails 3.0), it's possible to fetch gems
directly from specific branches in a git repository.  This is very handy
for contributors and people who need to make small fixes.

With this patch applied, it's possible to include the following line
in a Gemfile:

  gem 'mail', :git => 'git://github.com/mikel/mail.git'

Then you can install the latest master branch using bundler:

  bundle install

This patch changes two things about the development process.  First,
gems are now built using:

  rake build

...or:

  gem build mail.gemspec

Second, the generated gemfile now appears in the top-level directory,
and not in a pkg/ subdirectory.