Redid the folding to make it much much faster for long strings (the old one really crumbled with headers longer than 100,000 characters, which I happen to use in my setup with Sendgrid's SMTPAPI).
While I was at it, I decided to make it so that it would run longer stretches of encoded-words, and only one per line. The old implementation was really pessimistic in only taking at most 10 characters or up to the next white-space, encoding that (including all the overhead =?...?Q?...?=), and then repeating for the next 10-character stretch. This implementation tries to take as much as possible, given what it ends up looking like when it's encoded. The output is more compact and a lot prettier this way. Yay!

This version also fixes all the previous issues with extra white-space and an extra = at the end of the encoded-words.