Fixed charset issue with multipart messages.
When an add_file was called to add an attachment on a plain
(non-multipart) message, the call which was made to create a new "Part"
did not copy over the charset of the earlier plain-text message, but
instead copied over only the body. Later, it would always show a warning
message that the charset had to be detected, even though it was set
before.

Caveat: You need to set charset before the first call to add_file.