Enable TLS support under Ruby 1.8.x.
There needs to be a check for which version of ruby is being used
when trying to enable tls support.   Ruby 1.8 Net::SMTP does not
support the context object for enabling the TLS support, you
need to pass the OPEN::SSL::VERIFY related constant directly.

This patch checks for the version of Ruby and enables it as appropriate.