Split strictly on MIME boundary lines [Fixes #379]
Some mail generators ignore RFC 2046's "Boundary delimeters must not
appear within the encapsulated material" by using MIME boundaries for
inner parts that contain the MIME boundaries for outer parts.

To parse these emails as the author intended it is necessary to split
encapsulated content only on boundary delimeters that are on their own
line, instead of those that appear with suffixed content.