Fix parsing of content-type headers with " ;"
Before this change the header was accepted by the lexer, but due to an
ambiguity in the grammer it generated too many :param_val_e tokens.

This was because parameter (in the case that value was a quoted string)
expanded to:

   attr        -- value  --
  /    \     /              \
  token+ "=" '"' ... '"' FWS* FWS*

Ragel's minification algorithm then merged the FWS* into one rule, which
emitted a :param_val_e event for every whitespace token.

After this change, value can never end with whitespace, so the point at
which :param_val_e should be emitted is unambiguous.

The email I encountered this problem in had:

Content-Type: multipart/related;
              boundary=\"============_-840345472==_mr============\"
              ; type=\"text/html\""