Remove conditional dependencies from gemspec
Conditionals in a gemspec are only executed
when building the gem.

That is, when you run gem install, you get the built gem.
If the gem was built under CRuby 1.9.3, it would not specify
jruby-openssl or tlsmail as dependencies, no matter what
platform or version of Ruby you are running gem install on.

It only really makes sense to have them in the Gemfile for use in
development.