Use the "composition over inheritance" technique
Instead of inheriting from `Array`, let's create an attribute that is an
array. The current object doesn't fit with the Liskov substitution
principle as, for instance, #sort is broken on Rubinius.

This makes the `PartsList` class compatible with Rubinius as `undef map`
will remove the map method but some other Enumerable methods (e.g. sort)
rely on it.

Also add  fallback for Array's methods for the sake of backward
compatibility, we need to provide a fallback for missing method that may
be used in the existing libraries.

To avoid all the necessary boilerplate to provide such fallback, let's
rely on the standard library's delegate gem and its DelegateClass
method.