Fix Mail::Multibyte::Unicode::UnicodeDatabase#load
Temporarily define ActiveSupport::Multibyte so
Mail::Multibyte::Unicode::UnicodeDatabase#load works,
since it attempts to unmarshal an instance of
ActiveSupport::Multibyte::Unicode::Codepoint.

An better fix would be to change the unicode_tables.dat
format, so that ActiveSupport is not needed at all, but
I'm not sure of the other effects of that change.