Remove temporary definition of ActiveSupport::Multibyte
This removes the temporary definition, by hacking the unicode_tables.dat
file.  This is a ruby marshal format, so you can just change the
ActiveSupport::Multibyte::Unicode::Codepoint string to
Mail::Multibyte::Unicode::Codepoint, and then just adjust the length
of the class name, which is stored in the byte before the class name.