Stop collapsing adjacent encodings, fixing whitespace between encoded-words
Adjacent encoded-words with different character sets or encodings would
inadvertently leave valid separator characters (space \x20 or newline \x0A)
as an unencoded part. These characters, per the RFC1342 spec (page 3-4,
"Use of encoded-words in message headers") should not be displayed.

This fix is only for adjacent encoded-words and does not strip the
separator character (linear white-space or newline character) following
encoded words that are then followed by a "word", "text", "ctext", or
"special" (which per the spec should be stripped unless the separator is
a newline that comes at the end of the field.)

Closes #876