add use_args option to SendMail delivery method
if use_args is true then SendMail will invoke
the sendmail binary using execv style argument
passing instead of shell style argument passing.

this fixes issues where email addresses are
corrupted by shell escaping. this is an option
because it breaks compatibility when the
:arguments option is specified because the
:arguments option previously took a string of
arguments while when passing execv style we need
to take the option as an array of arguments.

this also breaks compatibility with the previous
behaviour because it now raises an error when
sendmail generates a non-zero exit code.

it also breaks compatibility with the previous
behaviour because use_args will not work with
older ruby versions. older ruby versions don't
support redirected stderr from popen and this is
probably why shell style argument passing was
originally used.