Work around email clients that may mistake a text attachment for the message body.
The default part sort order is

    [ "text/plain", "text/enriched", "text/html" ]

If a multipart email contains text and html parts as well as a text
attachment, the text attachment will often appear before the text body
in the message. In some mail clients this results in the attachment
being treated as the message body.

To avoid this ambiguity, sort attachments to the end of the parts list.

Signed-off-by: Jeremy Daer <jeremydaer@gmail.com>