Fix quote-handling regression from Ragel upgrade
Came across a regression in how quotes are handled in display names
after upgrading mysociety/alaveteli from Mail 2.5.4 to 2.6.x, similar to
issues #929 and #1017.

This commit fixes #929 and #1017.

I've included a spec similar to that demonstrated in #1017 and one that
mirrors the spec that highlighted the regression in mysociety/alaveteli.

Also included a spec to cover a quoted display name containing
backslash-escaped double quotes and parenthesis, with comments inside and
outside the address section.

I haven't included a spec for the failure example in #929, because I think that
`Mail::FromField` should expect the `Mail::Address` to be correct. The
spec for #1017 also covers the example in #929.