Replace mime-types with mini_mime to save memory
On boot, mime/types/columnar allocates **109k** objects and ends up
with a retained count of **31K** objects. This leaves objects that
need to be marked and swept every major GC and bloats ruby processes.

To circumvent, mini_mime was created. It uses the exact same database
as the full-fledged mime-types gem and is capable of only loading on
demand with a practical, safe and bound in-memory cache.

mini_mime allocates **398** objects on boot and only retain **62**
of them (due to rubygems inefficiency).

In table form:

| |boot allocations|boot retained|lookup|lookup uncached|
|-|-:|-:|-:|-:|
|mini_mime|398|62|641K/s|33K/s|
|mime-types|109796|31165|361K/s|361K/s|

The performance of mini_mime is pretty good: cached lookups are faster
than the mime-types gem and uncached lookups are 10x slower, which is
really not at huge issue considering the memory savings.