Remove unused methods in core_extensions
The `at`, `from`, `last` and `is_utf8?` patches to String in core_extensions
are not currently used by mail.

`to` is only used by `first` for Ruby < 2.0 but I kept both implementations
for consistency. `first` is only used in one location and only by Ruby < 2.0
but I also kept that to keep this PR focused on removed completely unused
patches.