Stop stripping header name prefixes from header values
Fixes mail with a legit "Subject: subject: foo" header. We'd split that
into a subject field with a "subject: foo" value. When we instantiate
the SubjectField and pass that value, we'd strip the leading "subject: "
from it again! So we'd end up with a "foo" subject.

This fixes the issue at the cost of some obscure backward compatibility.

1. If you instantiate internal header fields directly, bypassing
Mail::Field, then using them to parse raw headers will now fail.
This will now be treated as a literal value:

    Mail::ToField.new('To: foo@example.com')
    # => Mail::Field::IncompleteParseError: Mail::AddressList can not parse |To: foo@example.com|: Only able to parse up to "To: foo@example.com"

Instead, use Mail::Field.parse:

    Mail::Field.parse('To: foo@example.com')
    # => #<Mail::ToField …>

2. If you parse fields by instantiating Mail::Field.new with a raw header,
it'll now warn you with a deprecation error. That behavior will go away in
Mail 2.8.0.

    Mail::Field.new('To: foo@example.com')
    # => Passing an unparsed header field to Mail::Field.new is deprecated and will be removed in Mail 2.8.0. Use Mail::Field.parse instead.
    # => #<Mail::Field 0x7fa401026910 @charset="utf-8" @name="To" @value="foo">

Instead, use Mail::Field.parse:

    Mail::Field.parse('To: foo@example.com')
    # => #<Mail::ToField …>

Thanks to @grosser for 766 ❤️