Handle edge case when inline image attachment part is missing a filename
Fixes https://github.com/mikel/mail/issues/1329

First attempt at a fix: https://github.com/mikel/mail/pull/1332

In this second attempt, I tried to scope down the issue to where the issue appears most probable: inline images.

We do not add a default filename for the nameless attachment, rather just ensure that the message part is considered an attachment if it is an inline image.