Remove downcase! method from sanitize function in ContentTypeField
Fixes #1398

This change removes the sequence of code that downcases the Content-Type header value.
If I should believe the specs, it should serve no purpose. Only 2 specs break because
of the charset being returned in capital letters instead of downcased as before.

This fixes the problem is which encoding is broken by downcasing, such as the example
with base64 provided in #1398