Handle multiple singular headers when one of them is invalid
Singular header fields are supposed to be replaced when inserted, as there can
be only one of their type. However, if we happen to encounter an invalid
field among the duplicate ones, we'll fail to parse it and we'll keep it
as UnstructuredField. Since UnstructuredField aren't singular, we won't
replace the previous one we had for this header, they'll be stored
separately in the field list.

Then, later when fetching the field from the field list via
`#get_field`, we'll find the valid and invalid ones, and for singular
fields, this will raise a
```
undefined method `default' for #<Array:0x00007f40da2f2f70> Did you mean? to_default_s
```
since we'll be returning an array instead of a proper field that knows
how to handle multiple values.

This commit handles this case by checking, when getting multiple values,
whether one of them is singular and if this is the case, returning the
first one without errors, or if there's none without errors, the first
one.