Handle multiple invalid values for singular fields
This is a follow-up to 8ef377d6a2ca78aa5bd7f739813f5a0648482087. In that
commit we handled multiple values for singular fields when some of these
multiple values were valid. In this case we handle multiple values, all
of them invalid. In that case we can't trust any of the individual
instanced fields to respond to #singular?, since they won't be singular,
they will be UnstructuredFields. Then, when trying to fetch the value
via Mail::Message#field_name, we'll get a
```
undefined method `default' for #<Array:0x00007f40da2f2f70> Did you mean? to_default_s
```
error again, as we'd select all invalid fields for the singular field,
returning an array instead of a container that knows how to handle
multiple values.

To prevent this, we ask the class for the field name we're parsing
instead of relying on the field we might have failed to parse into the
right class.