Add Ruby 3.1 support
Add net-smtp as a dependency for Ruby 3.1 compatibility

It was recently removed from the default gems:

  - https://bugs.ruby-lang.org/issues/17873
  - https://github.com/ruby/ruby/pull/4530

Also add net-imap and net-pop to the gemspec

This dependencies introduces `strscan` as a subdependency, but JRuby
only has prerelease support for it. Make sure the prerelease is
properly picked up by our tests by adding it to the `Gemfile`.