Switch to truffleruby-head
The released version of `truffleruby` does not work with old versions of
Bundler. However, it's not possible to specify a custom version of
Bundler in truffleruby, because the `ruby/setup-ruby` will ignore that
in the case of `truffleruby-22.0` (the current release) That's in my
opinion an issue with `ruby/setup-ruby` because it should respect user's
choice when it's explicit. But it's not possible to change that because
it was an explicit choice of that library.

So as an alternative solution, use `truffleruby-head`, which apparently
does not have the issue.